Модуль 8, задание 1.
