+++ package.json
@@ -17,6 +17,7 @@
     "gulp-rename": "^2.0.0",
     "gulp-sourcemaps": "2.6.5",
     "gulp-svgstore": "^7.0.1",
+    "gulp-terser": "^1.4.0",
     "gulp-webp": "^4.0.1",
     "imagemin-webp": "^6.0.0",
     "stylelint": "13.6.1",
@@ -43,7 +44,9 @@
     "source/less/**/*.less"
   ],
   "engines": {
-    "node": "14.5.0"
+    "node": "12.18.3"
   },
-  "dependencies": {}
+  "dependencies": {
+    "gulp-htmlmin": "^5.0.1"
+  }
 }
