+++ source/catalog.html
@@ -1991,10 +1991,6 @@
       </div>
     </footer>
 
-    <script src="js/script-no-js.js"></script>
-    <script src="js/open-menu.js"></script>
-    <script src="js/open-filter.js"></script>
-    <script src="js/accordion.js"></script>
-    <script src="js/styleguide.js"></script>
+    <script src="js/script.min.js"></script>
   </body>
 </html>
