+++ source/form.html
@@ -1105,9 +1105,6 @@
       </div>
     </footer>
 
-    <script src="js/script-no-js.js"></script>
-    <script src="js/open-menu.js"></script>
-    <script src="js/open-choose.js"></script>
-    <script src="js/styleguide.js"></script>
+    <script src="js/script.min.js"></script>
   </body>
 </html>
