+++ source/index.html
@@ -862,10 +862,8 @@
       </div>
     </section>
 
-    <script src="js/script-no-js.js"></script>
-    <script src="js/open-menu.js"></script>
+    <script src="js/script.min.js"></script>
     <script src="https://api-maps.yandex.ru/2.1/?apikey=11ce66be-ec32-4218-8fea-46eb1c11e5be&lang=ru_RU" type="text/javascript"></script>
-    <script src="js/map.js"></script>
-    <script src="js/popup.js"></script>
+
   </body>
 </html>
