--- source/js/script-no-js.js
@@ -1,4 +1,0 @@
-let bodyPage = document.querySelector('body');
-
-// При успешной загрузке js снимается состояние без скрипта
-bodyPage.classList.remove('no-js');
